created userResource
